feed, comment 본인만 수정,삭제 가능

+++ src/main/java/stg/onyou/controller/CommentController.java
@@ -5,16 +5,16 @@ import io.swagger.annotations.Api;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
+import stg.onyou.exception.CustomException;
+import stg.onyou.exception.ErrorCode;
 import stg.onyou.model.entity.Comment;
-import stg.onyou.model.entity.Feed;
 import stg.onyou.model.network.Header;
-import stg.onyou.model.network.request.CommentUpdateRequest;
-import stg.onyou.model.network.response.CommentResponse;
 import stg.onyou.service.CommentService;
 import stg.onyou.service.FeedService;
+import stg.onyou.service.UserService;
 
-import java.util.List;
-import java.util.stream.Collectors;
+import javax.servlet.http.HttpServletRequest;
+import java.util.Objects;
 
 @Api(tags = {"Comment API Controller"})
 @Slf4j
@@ -24,6 +24,8 @@ public class CommentController {
     private final FeedService feedService;
     private final CommentService commentService;
 
+    private final UserService userService;
+
 //    @PutMapping("/api/comments/{id}")
 //    public Header<Object> updateComment(@PathVariable Long id,
 //                                        @RequestBody CommentUpdateRequest request) {
@@ -32,8 +34,14 @@ public class CommentController {
 //    }
 
     @DeleteMapping("/api/comments/{id}")
-    public Header<Object> deleteComment(@PathVariable Long id) {
-        commentService.deleteById(id);
-        return Header.OK("Comment 삭제 완료");
+    public Header<Object> deleteComment(@PathVariable Long id, HttpServletRequest httpServletRequest) {
+        Long userId = userService.getUserId(httpServletRequest);
+        Comment comment = commentService.findById(id);
+        if (Objects.equals(comment.getUser().getId(), userId)) {
+            commentService.deleteById(id);
+            return Header.OK("Comment 삭제 완료");
+        } else {
+            throw new CustomException(ErrorCode.NO_AUTH_DELETE_COMMENT);
+        }
     }
 }
